- Rev 108 - [2007/01/13] - Tsusai - Notes/Overview: Fixes and partial commits to character server for  zone connections. [2007/01/13] - Tsusai - CharacterServer : Added fZoneServerList, OnDisconnect Method,  updated SendCharaToMap to take on a connection and buffer parameter, Corrected error constant  numbers (was getting email error where could not happen error should be, and vice versa.  Please verify this change RaX). Added commented out code in SendcharaToMap. Fixed StatPoint  display in CreateChara.  Removed unneeded If connected check in ParseCharaServ. [2007/01/13] - Tsusai - ***GameDatabase: Updated GetChara to take on a JanSQLReleaseTable  hidden bool.  Assumes false. Updated any GetChara calls.  Updated LoadChara to attach  Account information. [2007/01/13] - Tsusai - CharacterServer, ZoneRecv: Updated any/all GetChara calls.
[2007/01/13] - Tsusai - Character: ClientVersion changed to byte (I need a negative number).
[2007/01/13] - Tsusai - Added ServerInfo root object, and updated CharacterServerInfo to
 descend from that.  Added ZoneServerInfo to do the same.
[2007/01/13] - Tsusai - PacketTypes: Added ZoneServerLink object.
-

+++ src/Classes/Character.pas
@@ -119,7 +119,7 @@ uses
 		Speed : word; //Not in MySQL...odd...
 		Account : TAccount;
 
-		ClientVersion : byte;
+		ClientVersion : Integer;
 		ClientInfo : TIdContext;
 
 		property DataChanged : boolean  read fDataChanged write SetSaveTime;
